Add the invocation request and response specification, plumb the enclave
Primary contribution of this patch is the invocation specification.
The specification defines the format of the JSON RPC that is made
into the interpreter to invoke methods on the contract. It includes
the format for the invocation request, invocation response, and the
execution environment.

Changes to the contract enclave enable collection of the information
necessary for the execution environment (specifically, providing
the hashes for the message and the code, and the name of the contract
class).

Signed-off-by: Mic Bowman <mic.bowman@intel.com>