Convert service groups to use a persistent file
Service group files loaded on demand has been one of the more painful
complexities for running contracts. This commit introduces a persistent
database comparable to the database used to store service endpoints. The
new groups database is stored at the location specified in the
['Service', 'GroupDatabaseFile'] configuration entry.

The service groups commands were all updated to accommodate the new
style. These are SIGNIFICANT changes. To ensure that correctness has
not been compromised, a new test was added for storage groups.

Finally, the persistent database required updating several other tests.
This change is, for the moment, a "functionally correct" change to the
tests (meaning that all tests now use the correct service groups commands)
but there is significant room for cleaning up the process.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>