Address a few more python setuptools upgrade issues
The current version of python setuptools no longer puts binary files
in the same name directory tree as the python libraries. As a result,
the eservice and pservice enclaves could not be found and loaded.

This commit addresses the problem with several changes:

1) the names of the enclave libraries are now unique
(libpdo-eservice-enclave.signed.so and
libpdo-pservice-enclave.signed.so)

2) the enclave libraries are installed into the PDO_HOME directory
tree (in the lib directory). This directory is now created when the
rest of the enviroment is created.

3) the search path used by the pservice and eservice python
modules reflect the new location.

4) cachetools is explicitly installed

Signed-off-by: Mic Bowman <mic.bowman@intel.com>